+++ public/js/components/instruction-slides.js
