+++ public/js/components/submit-location-timestamp.js
